% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateDistance.R
\name{CalculateDistance}
\alias{CalculateDistance}
\title{Obliczanie dystansu obiektów geograficznych od centrum Krakowa}
\usage{
CalculateDistance(data, longtitude, latitude)
}
\arguments{
\item{data}{Obiekt typu data.frame zawierający dane obiektów, w tym
położenie geografincze (tj. szerokość geograficzną - zmienna
\code{latitude} i długość geograficzna - zmienna \code{longtitude})}

\item{longtitude}{Wektor wartości typu numeric, reprezentujących długość
geograficzną wyrażoną w stopniach dziesiętnych, które wraz z
korespondującymi z nimi rekordami zmiennej \code{longtitude} dają pełną
infomrację o współrzędnych geograficznych. Może przyjmować wartości w
zakresie od -180 do 180.}

\item{latitude}{Wektor wartości typu numeric reprezentujących szerokość
geograficzną wyrażoną w stopniach dziesiętnych, które wraz z
korespondującymi z nimi rekordami zmiennej \code{latitude} dają pełną
infomrację o współrzędnych geograficznych. Może przyjmować wartości
w zakresie od -90 do 90.}
}
\value{
Wektor wartości typu numeric, reprezetujących dystans obiektów
(w km) od centurm Krakowa (tj. Rynku Głównego).
}
\description{
Funkcja oblicza dystans (w km) od Centrum Krakowa (tj. Rynku Głównego)
}
\examples{
\dontrun{
#Uzupełenianie bazy obiektów turystycznych Krakowa i okolic o informację
#dot.ich odległości od Rynku Głównego
new.df <- df[, distance:=CalculateDistance(df,"lng","lat")]}
}
