% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDescription.R
\docType{data}
\name{df}
\alias{df}
\title{Zbiór danych dot. głównych obiektów turystycznych Krakowa i okolic}
\format{Ramka danych (data.frame) \code{df} w formacie data.table
zawierająca 119 rekordów i 4 zmienne:
\describe{
  \item{\code{name}}{nazwa obiektu, w języku angielskim}
  \item{\code{lat}}{szerokość geograficzna obiektu}
  \item{\code{lng}}{długość geograficzna obiektu}
  \item{\code{type}}{typ obiektu - kategoria oznaczona subiektywnie}
}}
\usage{
data(data)
}
\description{
Zawiera dane geolokalizacyjne subiektywnie wybranych obiektów
turystycznych
}
\examples{
\dontrun{
data(data)
str(df)}

}
\keyword{datasets}
