% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map.R
\name{Map}
\alias{Map}
\title{Wizualizacja geolokalizacji obiektów}
\usage{
Map(data = NULL, longtitude = NULL, latitude = NULL, name.type = NULL)
}
\arguments{
\item{...}{Arguments passed on to \code{SelectTypeDistance}
\describe{
  \item{name.type}{Nazwa zmiennej w ramce danych \code{data} zawierająca
kategorie typu obiektu. Zmienna typu character.}
  \item{data}{Obiekt typu data.frame zawierający dane obiektów, w tym
położenie geografincze (tj. szerokość geograficzną - zmienna
\code{latitude} i długość geograficzna - zmienna \code{longtitude})}
  \item{latitude}{Wektor wartości typu numeric reprezentujących szerokość
geograficzną wyrażoną w stopniach dziesiętnych, które wraz z
korespondującymi z nimi rekordami zmiennej \code{latitude} dają pełną
infomrację o współrzędnych geograficznych. Może przyjmować wartości
w zakresie od -90 do 90.}
  \item{longtitude}{Wektor wartości typu numeric, reprezentujących długość
geograficzną wyrażoną w stopniach dziesiętnych, które wraz z
korespondującymi z nimi rekordami zmiennej \code{longtitude} dają pełną
infomrację o współrzędnych geograficznych. Może przyjmować wartości w
zakresie od -180 do 180.}
}}
}
\value{
Tożsame z parametrem Value w [leaflet::leaflet]
}
\description{
Plasowanie na mapie obiektów, których dane geolokalizacyjne zamieszczone
w ramce danych wraz z oznaczeniem ich rodzaju/ kategorii.
}
\examples{
#Wizualizacja na mapie obiektów kulutry judaizmu w Krakowie położonych nie
#dalej niż 3km od Rynku Głównego
\dontrun{
new.df <- SelectTypeDistance(data = df, latitude = "lat",
longtitude = "lng", name.type = "type", sel.type = "Juish Culture",
sel.dist = 3)
Map(data = new.df, longtitude = "lng", latitude = "lat",
name.type = "type")}
}
