% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectTypeDistance.R
\name{SelectTypeDistance}
\alias{SelectTypeDistance}
\title{Selekcja obiektów}
\usage{
SelectTypeDistance(data = NULL, latitude = NULL, longtitude = NULL,
  name.type = NULL, sel.type = NULL, sel.dist = 60)
}
\arguments{
\item{data}{Obiekt typu data.frame zawierający dane obiektów, w tym
położenie geografincze (tj. szerokość geograficzną - zmienna
\code{latitude} i długość geograficzna - zmienna \code{longtitude})}

\item{latitude}{Wektor wartości typu numeric reprezentujących szerokość
geograficzną wyrażoną w stopniach dziesiętnych, które wraz z
korespondującymi z nimi rekordami zmiennej \code{latitude} dają pełną
infomrację o współrzędnych geograficznych. Może przyjmować wartości
w zakresie od -90 do 90.}

\item{longtitude}{Wektor wartości typu numeric, reprezentujących długość
geograficzną wyrażoną w stopniach dziesiętnych, które wraz z
korespondującymi z nimi rekordami zmiennej \code{longtitude} dają pełną
infomrację o współrzędnych geograficznych. Może przyjmować wartości w
zakresie od -180 do 180.}

\item{name.type}{Nazwa zmiennej w ramce danych \code{data} zawierająca
kategorie typu obiektu. Zmienna typu character.}

\item{sel.type}{Wybrana kategoria typu obiektu (mieszcząca się w zmiennej
\code{name.type}), w oparciu o którą będzie filtrowana ramka danych
\code{data}.}

\item{sel.dist}{Wartość numeryczna (numeric). Kryterium wyboru obiektów
określające maksymalna odległość obiektu (w km) od centrum Krakowa. Domyślnie
wynosi 60km.}
}
\value{
Obiekty typu data.frame.
}
\description{
Pozwala na selekcjonowane obiektów ramki danych w oparciu o kryterium typu
obiektu oraz odległości, która dzieli je od centrum Krakowa (tj. Rynku
Głównego).
}
\examples{
#Wybór z bazy obiektów turystycznych Krakowa wyłącznie muzeów i kościołów
#położonych nie dalej niż 2km od Rynku Głównego
\dontrun{
new.df<-SelectTypeDistance(data = df, latitude = "lat",
longtitude = "lng", name.type = "type",
sel.type = c("Churches", "Museums"), sel.dist = 2)
}
}
