% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SettingWidth.R
\name{SettingWidth}
\alias{SettingWidth}
\title{Ustalanie parametru width dla funkcji geom_bar}
\usage{
SettingWidth(n = NULL)
}
\arguments{
\item{n}{Wartość typu numeric reprezentująca liczbę kategorii, które
zostaną oznaczone na wykresie słupkowym.}
}
\value{
Wartość typu numeric.
}
\description{
Funkcja pozwoli wyznaczyć optymalną szerokość słupków na wykresie w
zalezności od liczby kategorii zmiennej
}
\examples{
\dontrun{
n <- length(unique(df$type))
SettingWidth(n)}
}
