% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShowHowMany.R
\name{ShowHowMany}
\alias{ShowHowMany}
\title{Wizualizacja liczby obiektów}
\usage{
ShowHowMany(data = NULL, name.type = NULL)
}
\arguments{
\item{data}{Obiekt typu data.frame zawierający przynajmniej jedną zmienną
typu character, reprezentującą kategorie typu obiektu.}

\item{name.type}{Nazwa zmiennej w ramce danych \code{data} zawierająca
kategorie typu obiektu. Zmienna typu character.}
}
\value{
Wartość typu numeric, do wykorzystania jako parametr width
funkncji geom_bar.
}
\description{
Funkcja generuje wykres słupkowy prezentacjący liczbę poszczególnych rodzajów
obiektów znajdujących się w ramce danych
}
\examples{
\dontrun{
new.df <- SelectTypeDistance(data = df, latitude = "lat", longtitude = "lng",
name.type = "type", sel.type = c("Streets and Squares", "Parks and Nature"),
sel.dist = 20)
ShowHowMany(data = new.df, name.type = "type")}
}
